chore: move connection to an abstract base class
