+++ pynamodb/connection/base.py
@@ -3,8 +3,7 @@ Lowest level connection
 """
 import logging
 import uuid
-from threading import local
-from typing import Any, Dict, List, Mapping, Optional, Sequence, cast
+from typing import Any, Dict, List, Optional, Sequence, cast
 
 import botocore.client
 import botocore.exceptions
@@ -13,28 +12,22 @@ from botocore.exceptions import BotoCoreError
 from botocore.session import get_session
 
 from pynamodb.connection._botocore_private import BotocoreBaseClientPrivate
-from pynamodb._util import bin_decode_attr
+from pynamodb.connection.skeleton import ConnectionSkeleton
 from pynamodb.constants import (
-    RETURN_CONSUMED_CAPACITY_VALUES, RETURN_ITEM_COLL_METRICS_VALUES,
-    RETURN_ITEM_COLL_METRICS, RETURN_CONSUMED_CAPACITY, RETURN_VALUES_VALUES,
-    EXCLUSIVE_START_KEY, SCAN_INDEX_FORWARD, ATTR_DEFINITIONS,
+    RETURN_CONSUMED_CAPACITY, EXCLUSIVE_START_KEY, SCAN_INDEX_FORWARD, ATTR_DEFINITIONS,
     BATCH_WRITE_ITEM, CONSISTENT_READ, DESCRIBE_TABLE, KEY_CONDITION_EXPRESSION,
-    BATCH_GET_ITEM, DELETE_REQUEST, SELECT_VALUES, RETURN_VALUES, REQUEST_ITEMS,
+    BATCH_GET_ITEM, DELETE_REQUEST, SELECT_VALUES, REQUEST_ITEMS,
     PROJECTION_EXPRESSION, SERVICE_NAME, DELETE_ITEM, PUT_REQUEST, UPDATE_ITEM, TABLE_NAME,
     INDEX_NAME, KEY_SCHEMA, ATTR_NAME, ATTR_TYPE, TABLE_KEY, KEY_TYPE, GET_ITEM, UPDATE,
     PUT_ITEM, SELECT, LIMIT, QUERY, SCAN, ITEM, LOCAL_SECONDARY_INDEXES,
     KEYS, KEY, SEGMENT, TOTAL_SEGMENTS, CREATE_TABLE, PROVISIONED_THROUGHPUT, READ_CAPACITY_UNITS,
     WRITE_CAPACITY_UNITS, GLOBAL_SECONDARY_INDEXES, PROJECTION, EXCLUSIVE_START_TABLE_NAME, TOTAL,
-    DELETE_TABLE, UPDATE_TABLE, LIST_TABLES, GLOBAL_SECONDARY_INDEX_UPDATES, ATTRIBUTES,
-    CONSUMED_CAPACITY, CAPACITY_UNITS, ATTRIBUTE_TYPES,
-    ITEMS, LAST_EVALUATED_KEY, RESPONSES, UNPROCESSED_KEYS,
-    UNPROCESSED_ITEMS, STREAM_SPECIFICATION, STREAM_VIEW_TYPE, STREAM_ENABLED,
+    DELETE_TABLE, UPDATE_TABLE, LIST_TABLES, GLOBAL_SECONDARY_INDEX_UPDATES, CONSUMED_CAPACITY, CAPACITY_UNITS, ATTRIBUTE_TYPES,
+    STREAM_SPECIFICATION, STREAM_VIEW_TYPE, STREAM_ENABLED,
     EXPRESSION_ATTRIBUTE_NAMES, EXPRESSION_ATTRIBUTE_VALUES,
-    CONDITION_EXPRESSION, FILTER_EXPRESSION,
-    TRANSACT_WRITE_ITEMS, TRANSACT_GET_ITEMS, CLIENT_REQUEST_TOKEN, TRANSACT_ITEMS, TRANSACT_CONDITION_CHECK,
-    TRANSACT_GET, TRANSACT_PUT, TRANSACT_DELETE, TRANSACT_UPDATE, UPDATE_EXPRESSION,
-    RETURN_VALUES_ON_CONDITION_FAILURE_VALUES, RETURN_VALUES_ON_CONDITION_FAILURE,
-    AVAILABLE_BILLING_MODES, DEFAULT_BILLING_MODE, BILLING_MODE, PAY_PER_REQUEST_BILLING_MODE,
+    FILTER_EXPRESSION,
+    TRANSACT_WRITE_ITEMS, TRANSACT_GET_ITEMS, TRANSACT_ITEMS, TRANSACT_CONDITION_CHECK,
+    TRANSACT_GET, TRANSACT_PUT, TRANSACT_DELETE, TRANSACT_UPDATE, AVAILABLE_BILLING_MODES, DEFAULT_BILLING_MODE, BILLING_MODE, PAY_PER_REQUEST_BILLING_MODE,
     PROVISIONED_BILLING_MODE, TIME_TO_LIVE_SPECIFICATION, ENABLED, UPDATE_TIME_TO_LIVE, TAGS, VALUE
 )
 from pynamodb.exceptions import (
@@ -45,8 +38,7 @@ from pynamodb.exceptions import (
 from pynamodb.expressions.condition import Condition
 from pynamodb.expressions.operand import Path
 from pynamodb.expressions.projection import create_projection_expression
-from pynamodb.expressions.update import Action, Update
-from pynamodb.settings import get_settings_value
+from pynamodb.expressions.update import Action
 from pynamodb.signals import pre_dynamodb_send, post_dynamodb_send
 from pynamodb.types import HASH, RANGE
 
@@ -236,63 +228,10 @@ class MetaTable(object):
             }
 
 
-class Connection(object):
+class Connection(ConnectionSkeleton):
     """
     A higher level abstraction over botocore
     """
-
-    def __init__(self,
-                 region: Optional[str] = None,
-                 host: Optional[str] = None,
-                 read_timeout_seconds: Optional[float] = None,
-                 connect_timeout_seconds: Optional[float] = None,
-                 max_retry_attempts: Optional[int] = None,
-                 max_pool_connections: Optional[int] = None,
-                 extra_headers: Optional[Mapping[str, str]] = None,
-                 aws_access_key_id: Optional[str] = None,
-                 aws_secret_access_key: Optional[str] = None,
-                 aws_session_token: Optional[str] = None):
-        self._tables: Dict[str, MetaTable] = {}
-        self.host = host
-        self._local = local()
-        self._client: Optional[BotocoreBaseClientPrivate] = None
-        self._convert_to_request_dict__endpoint_url = False
-        if region:
-            self.region = region
-        else:
-            self.region = get_settings_value('region')
-
-        if connect_timeout_seconds is not None:
-            self._connect_timeout_seconds = connect_timeout_seconds
-        else:
-            self._connect_timeout_seconds = get_settings_value('connect_timeout_seconds')
-
-        if read_timeout_seconds is not None:
-            self._read_timeout_seconds = read_timeout_seconds
-        else:
-            self._read_timeout_seconds = get_settings_value('read_timeout_seconds')
-
-        if max_retry_attempts is not None:
-            self._max_retry_attempts_exception = max_retry_attempts
-        else:
-            self._max_retry_attempts_exception = get_settings_value('max_retry_attempts')
-
-        if max_pool_connections is not None:
-            self._max_pool_connections = max_pool_connections
-        else:
-            self._max_pool_connections = get_settings_value('max_pool_connections')
-
-        if extra_headers is not None:
-            self._extra_headers = extra_headers
-        else:
-            self._extra_headers = get_settings_value('extra_headers')
-
-        self._aws_access_key_id = aws_access_key_id
-        self._aws_secret_access_key = aws_secret_access_key
-        self._aws_session_token = aws_session_token
-
-    def __repr__(self) -> str:
-        return "Connection<{}>".format(self.client.meta.endpoint_url)
 
     def dispatch(self, operation_name: str, operation_kwargs: Dict) -> Dict:
         """
@@ -413,23 +352,6 @@ class Connection(object):
 
             self._client.meta.events.register_first('before-send.*.*', self._before_send)
         return self._client
-
-    def add_meta_table(self, meta_table: MetaTable) -> None:
-        """
-        Adds information about the table's schema.
-        """
-        if meta_table.table_name in self._tables:
-            raise ValueError(f"Meta-table for '{meta_table.table_name}' already added")
-        self._tables[meta_table.table_name] = meta_table
-
-    def get_meta_table(self, table_name: str) -> MetaTable:
-        """
-        Returns information about the table's schema.
-        """
-        try:
-            return self._tables[table_name]
-        except KeyError:
-            raise TableError(f"Meta-table for '{table_name}' not initialized") from None
 
     def create_table(
         self,
@@ -631,7 +553,7 @@ class Connection(object):
                 meta_table = MetaTable(table_data)
                 if meta_table.table_name not in self._tables:
                     self.add_meta_table(meta_table)
-            return table_data
+            return cast(Dict, table_data)
         except BotoCoreError as e:
             raise TableError("Unable to describe table: {}".format(e), e)
         except ClientError as e:
@@ -639,186 +561,6 @@ class Connection(object):
                 raise TableDoesNotExist(e.response['Error']['Message'])
             else:
                 raise
-
-    def get_item_attribute_map(
-        self,
-        table_name: str,
-        attributes: Any,
-        item_key: str = ITEM,
-        pythonic_key: bool = True,
-    ) -> Dict:
-        """
-        Builds up a dynamodb compatible AttributeValue map
-        """
-        tbl = self.get_meta_table(table_name)
-        if tbl is None:
-            raise TableError("No such table {}".format(table_name))
-        return tbl.get_item_attribute_map(
-            attributes,
-            item_key=item_key,
-            pythonic_key=pythonic_key)
-
-    def parse_attribute(
-        self,
-        attribute: Any,
-        return_type: bool = False
-    ) -> Any:
-        """
-        Returns the attribute value, where the attribute can be
-        a raw attribute value, or a dictionary containing the type:
-        {'S': 'String value'}
-        """
-        if isinstance(attribute, dict):
-            for key in ATTRIBUTE_TYPES:
-                if key in attribute:
-                    if return_type:
-                        return key, attribute.get(key)
-                    return attribute.get(key)
-            raise ValueError("Invalid attribute supplied: {}".format(attribute))
-        else:
-            if return_type:
-                return None, attribute
-            return attribute
-
-    def get_attribute_type(
-        self,
-        table_name: str,
-        attribute_name: str,
-        value: Optional[Any] = None
-    ) -> str:
-        """
-        Returns the proper attribute type for a given attribute name
-        :param value: The attribute value an be supplied just in case the type is already included
-        """
-        tbl = self.get_meta_table(table_name)
-        if tbl is None:
-            raise TableError("No such table {}".format(table_name))
-        return tbl.get_attribute_type(attribute_name, value=value)
-
-    def get_identifier_map(
-        self,
-        table_name: str,
-        hash_key: str,
-        range_key: Optional[str] = None,
-        key: str = KEY
-    ) -> Dict:
-        """
-        Builds the identifier map that is common to several operations
-        """
-        tbl = self.get_meta_table(table_name)
-        if tbl is None:
-            raise TableError("No such table {}".format(table_name))
-        return tbl.get_identifier_map(hash_key, range_key=range_key, key=key)
-
-    def get_consumed_capacity_map(self, return_consumed_capacity: str) -> Dict:
-        """
-        Builds the consumed capacity map that is common to several operations
-        """
-        if return_consumed_capacity.upper() not in RETURN_CONSUMED_CAPACITY_VALUES:
-            raise ValueError("{} must be one of {}".format(RETURN_ITEM_COLL_METRICS, RETURN_CONSUMED_CAPACITY_VALUES))
-        return {
-            RETURN_CONSUMED_CAPACITY: str(return_consumed_capacity).upper()
-        }
-
-    def get_return_values_map(self, return_values: str) -> Dict:
-        """
-        Builds the return values map that is common to several operations
-        """
-        if return_values.upper() not in RETURN_VALUES_VALUES:
-            raise ValueError("{} must be one of {}".format(RETURN_VALUES, RETURN_VALUES_VALUES))
-        return {
-            RETURN_VALUES: str(return_values).upper()
-        }
-
-    def get_return_values_on_condition_failure_map(
-        self,
-        return_values_on_condition_failure: str
-    ) -> Dict:
-        """
-        Builds the return values map that is common to several operations
-        """
-        if return_values_on_condition_failure.upper() not in RETURN_VALUES_VALUES:
-            raise ValueError("{} must be one of {}".format(
-                RETURN_VALUES_ON_CONDITION_FAILURE,
-                RETURN_VALUES_ON_CONDITION_FAILURE_VALUES
-            ))
-        return {
-            RETURN_VALUES_ON_CONDITION_FAILURE: str(return_values_on_condition_failure).upper()
-        }
-
-    def get_item_collection_map(self, return_item_collection_metrics: str) -> Dict:
-        """
-        Builds the item collection map
-        """
-        if return_item_collection_metrics.upper() not in RETURN_ITEM_COLL_METRICS_VALUES:
-            raise ValueError("{} must be one of {}".format(RETURN_ITEM_COLL_METRICS, RETURN_ITEM_COLL_METRICS_VALUES))
-        return {
-            RETURN_ITEM_COLL_METRICS: str(return_item_collection_metrics).upper()
-        }
-
-    def get_exclusive_start_key_map(self, table_name: str, exclusive_start_key: str) -> Dict:
-        """
-        Builds the exclusive start key attribute map
-        """
-        tbl = self.get_meta_table(table_name)
-        if tbl is None:
-            raise TableError("No such table {}".format(table_name))
-        return tbl.get_exclusive_start_key_map(exclusive_start_key)
-
-    def get_operation_kwargs(
-        self,
-        table_name: str,
-        hash_key: str,
-        range_key: Optional[str] = None,
-        key: str = KEY,
-        attributes: Optional[Any] = None,
-        attributes_to_get: Optional[Any] = None,
-        actions: Optional[Sequence[Action]] = None,
-        condition: Optional[Condition] = None,
-        consistent_read: Optional[bool] = None,
-        return_values: Optional[str] = None,
-        return_consumed_capacity: Optional[str] = None,
-        return_item_collection_metrics: Optional[str] = None,
-        return_values_on_condition_failure: Optional[str] = None
-    ) -> Dict:
-        self._check_condition('condition', condition)
-
-        operation_kwargs: Dict[str, Any] = {}
-        name_placeholders: Dict[str, str]  = {}
-        expression_attribute_values: Dict[str, Any] = {}
-
-        operation_kwargs[TABLE_NAME] = table_name
-        operation_kwargs.update(self.get_identifier_map(table_name, hash_key, range_key, key=key))
-        if attributes and operation_kwargs.get(ITEM) is not None:
-            attrs = self.get_item_attribute_map(table_name, attributes)
-            operation_kwargs[ITEM].update(attrs[ITEM])
-        if attributes_to_get is not None:
-            projection_expression = create_projection_expression(attributes_to_get, name_placeholders)
-            operation_kwargs[PROJECTION_EXPRESSION] = projection_expression
-        if condition is not None:
-            condition_expression = condition.serialize(name_placeholders, expression_attribute_values)
-            operation_kwargs[CONDITION_EXPRESSION] = condition_expression
-        if consistent_read is not None:
-            operation_kwargs[CONSISTENT_READ] = consistent_read
-        if return_values is not None:
-            operation_kwargs.update(self.get_return_values_map(return_values))
-        if return_values_on_condition_failure is not None:
-            operation_kwargs.update(self.get_return_values_on_condition_failure_map(return_values_on_condition_failure))
-        if return_consumed_capacity is not None:
-            operation_kwargs.update(self.get_consumed_capacity_map(return_consumed_capacity))
-        if return_item_collection_metrics is not None:
-            operation_kwargs.update(self.get_item_collection_map(return_item_collection_metrics))
-        if actions is not None:
-            update_expression = Update(*actions)
-            operation_kwargs[UPDATE_EXPRESSION] = update_expression.serialize(
-                name_placeholders,
-                expression_attribute_values
-            )
-        if name_placeholders:
-            operation_kwargs[EXPRESSION_ATTRIBUTE_NAMES] = self._reverse_dict(name_placeholders)
-        if expression_attribute_values:
-            operation_kwargs[EXPRESSION_ATTRIBUTE_VALUES] = expression_attribute_values
-        return operation_kwargs
 
     def delete_item(
         self,
@@ -908,22 +650,6 @@ class Connection(object):
             return self.dispatch(PUT_ITEM, operation_kwargs)
         except BOTOCORE_EXCEPTIONS as e:
             raise PutError("Failed to put item: {}".format(e), e)
-
-    def _get_transact_operation_kwargs(
-        self,
-        client_request_token: Optional[str] = None,
-        return_consumed_capacity: Optional[str] = None,
-        return_item_collection_metrics: Optional[str] = None
-    ) -> Dict:
-        operation_kwargs = {}
-        if client_request_token is not None:
-            operation_kwargs[CLIENT_REQUEST_TOKEN] = client_request_token
-        if return_consumed_capacity is not None:
-            operation_kwargs.update(self.get_consumed_capacity_map(return_consumed_capacity))
-        if return_item_collection_metrics is not None:
-            operation_kwargs.update(self.get_item_collection_map(return_item_collection_metrics))
-
-        return operation_kwargs
 
     def transact_write_items(
         self,
@@ -1211,12 +937,3 @@ class Connection(object):
             return self.dispatch(QUERY, operation_kwargs)
         except BOTOCORE_EXCEPTIONS as e:
             raise QueryError("Failed to query items: {}".format(e), e)
-
-    def _check_condition(self, name, condition):
-        if condition is not None:
-            if not isinstance(condition, Condition):
-                raise ValueError("'{}' must be an instance of Condition".format(name))
-
-    @staticmethod
-    def _reverse_dict(d):
-        return {v: k for k, v in d.items()}
