+++ pyproject.toml
@@ -28,6 +28,7 @@ pytest-asyncio = "^0.25.3"
 [tool.poetry.group.dev.dependencies]
 types-aioboto3 = {extras = ["dynamodb"], version = "^13.4.0"}
 mypy = "^1.15.0"
+ruff = "^0.9.6"
 
 [build-system]
 requires = ["poetry-core"]
